TranslationService/EN2AR/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.115.6
pydantic==2.10.3
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
uvicorn==0.32.1
psycopg2-binary
